+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Undo deprecation of `NewProviderGroup`.
+
+### Fixed
+- Handle empty sources correctly in `NewProviderGroup`.
+
 ## 1.2.1 - 2018-07-06
 ### Fixed
 - Handle empty sources and top-level nulls correctly.